pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pixelterm"
version = "0.1.0"
description = "A modern terminal image browser based on chafa"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "chafapy>=0.1.0",
    "Pillow>=9.0.0",
]

[project.scripts]
pixelterm = "pixelterm:main"